CCP-1729: fixed some jsDoc. Moved and renamed moutScript function. Adjusted configuration
