+++ frontend/constants.js
@@ -9,6 +9,3 @@ export const SMILE_DIGEST_PIPELINE = 'shopgate-project.smile-io.getCustomerDiges
 
 // Redux namespace
 export const REDUX_NAMESPACE_SMILE_DIGEST_DATA = '@shopgate-project/smile-io/smileDataDigest';
-
-// Misc
-export const SWEETTOOTH_INIT_ENDPOINT = 'https://cdn.sweettooth.io/v1/storefront_js/init';
