--- frontend/helpers/getConfig.js
@@ -1,23 +1,0 @@
-import { logger } from '@shopgate/pwa-core/helpers';
-
-let cached;
-
-/**
- * Gets extension config or empty object on failure.
- * @returns {Object}
- */
-function getConfig() {
-  if (!cached) {
-    try {
-      // eslint-disable-next-line global-require
-      cached = require('../config');
-    } catch (e) {
-      logger.error('Could not read config.', e);
-      cached = {};
-    }
-  }
-
-  return cached;
-}
-
-export default getConfig;
