+++ frontend/portals/SmileContainer/index.jsx
@@ -1,25 +1,23 @@
 import React, { Fragment, useEffect } from 'react';
 import PropTypes from 'prop-types';
-import getConfig from '../../helpers/getConfig';
 import connect from './connector';
-
-const { smileChannelApiKey } = getConfig();
+import { smileChannelApiKey } from '../../config';
 
 /**
  * SmileContainer component
  * @param {string} digest Smile digest data
  * @param {string} externalCustomerId External Customer Id
- * @param {Function} mountScript Mount script function
+ * @param {Function} initializeScript Initialize Smile.io script function
  * @return {JSX}
  */
-const SmileContainer = ({ digest, externalCustomerId, mountScript }) => {
+const SmileContainer = ({ digest, externalCustomerId, initializeScript }) => {
   useEffect(() => {
     if (!(digest && externalCustomerId)) {
       return;
     }
 
-    mountScript(externalCustomerId, digest);
-  }, [digest]);
+    initializeScript(externalCustomerId, digest);
+  }, [digest, externalCustomerId]);
 
   if (!(digest && externalCustomerId)) {
     return null;
@@ -41,13 +39,13 @@ const SmileContainer = ({ digest, externalCustomerId, mountScript }) => {
 SmileContainer.propTypes = {
   digest: PropTypes.string,
   externalCustomerId: PropTypes.string,
-  mountScript: PropTypes.func,
+  initializeScript: PropTypes.func,
 };
 
 SmileContainer.defaultProps = {
   digest: null,
   externalCustomerId: null,
-  mountScript: () => {},
+  initializeScript: () => {},
 };
 
 export default connect(SmileContainer);
